@torzdf 2019-07-09
This looks fine. I think remove_alignment is still hanging around from before filter_hashes was implemented (which does the same job)